Refactored code, removed some print and solved a problem with the Double to Float.

+++ src/main/scala/domain/distribution/value/DistributionFlowNumberScala.scala
@@ -4,6 +4,6 @@ import domain.in.distribution.InputDistributionScala
 import domain.transform.calculations.CalculationsScala
 import domain.value.generator.ValueGeneratorScala
 
-final class DistributionFlowNumberScala(override val dataGenerator: ValueGeneratorScala[Float, CalculationsScala[Float]],
+final class DistributionFlowNumberScala(override val dataGenerator: ValueGeneratorScala[Double, CalculationsScala[Double]],
                                         override val inputDistribution: Seq[InputDistributionScala])
-  extends DistributionFlowTScala[Float, CalculationsScala[Float]](dataGenerator, inputDistribution)
+  extends DistributionFlowTScala[Double, CalculationsScala[Double]](dataGenerator, inputDistribution)
